requirements.txt
boto3
pefile
psycopg2-binary
pyspark
